requirements.txt
affine==2.3.0
attrs==20.2.0
certifi==2020.6.20
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
Fiona==1.8.17
fire==0.3.1
geopandas==0.8.1
munch==2.5.0
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
rasterio==1.1.8
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
termcolor==1.1.0
tqdm==4.50.2
